pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pictovector"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Uroj Ashfa", github = "https://github.com/UroojAshfa", role = "Frontend Development"},
    {name = "Maran", github = "https://github.com/IlamaranMagesh", role = "Backend Development"}
]
dependencies= [
    "torch ~= 2.2.0",
    "protobuf ~= 6.33.1",
    "pillow ~= 12.0.0",
    "transformers ~= 4.57.1",
    "qdrant-client ~= 1.15.1",
    "python-dotenv ~= 1.2.1",
    "Flask ~= 3.1.2",
    "vercel ~= 0.3.6",
    "Werkzeug ~= 3.1.3",
    "datasets ~= 4.4.1",
    "requests ~= 2.32.5",
]

[project.scripts]
runapp-picto = "pictovector.index:app"

[project.urls]
Repository = "https://github.com/IlamaranMagesh/pictovector"
